In class definition interface ComponentPopulate now has a type parameter.
Implemented method handle_setResult in the AbstractMercuryReference anonymous reference constructed in method newInlinePanelForChoice().

In method handle_choiceItem(...) changed call to panel.data.setValue(...) to panel.setData().  Now the components are updated with the new value.

In method valueChanged(...) changed call to bpi.data.setValue(...) to bpi.setData().  Now the components in the shown panel are updated with the new value.

Implemented method commitValue() of interface ComponentPopulate